32-bit compile fix

+++ plink_common.h
@@ -2209,7 +2209,7 @@ static inline void unroll_zero_incr_8_32(__m128i* acc8, __m128i* acc32, uint32_t
 #else
 static inline void unroll_incr_1_4(const uintptr_t* acc1, uintptr_t acc4, uint32_t acc1_word_ct) {
   uint32_t widx;
-  uint32_t loader;
+  uintptr_t loader;
   for (widx = 0; widx < acc1_word_ct; widx++) {
     loader = *acc1++;
     *acc4 += loader & 0x11111111U;
@@ -2228,7 +2228,7 @@ static inline void unroll_incr_1_4(const uintptr_t* acc1, uintptr_t acc4, uint32
 
 static inline void unroll_incr_4_8(const uintptr_t* acc4, uintptr_t acc8, uint32_t acc4_word_ct) {
   uint32_t widx;
-  uint32_t loader;
+  uintptr_t loader;
   for (widx = 0; widx < acc4_word_ct; widx++) {
     loader = *acc4++;
     *acc8 += loader & 0x0f0f0f0fU;
@@ -2241,7 +2241,7 @@ static inline void unroll_incr_4_8(const uintptr_t* acc4, uintptr_t acc8, uint32
 
 static inline void unroll_zero_incr_4_8(uintptr_t* acc4, uintptr_t acc8, uint32_t acc4_word_ct) {
   uint32_t widx;
-  uint32_t loader;
+  uintptr_t loader;
   for (widx = 0; widx < acc4_word_ct; widx++) {
     loader = *acc4;
     *acc4++ = 0;
@@ -2255,7 +2255,7 @@ static inline void unroll_zero_incr_4_8(uintptr_t* acc4, uintptr_t acc8, uint32_
 
 static inline void unroll_incr_8_32(const uintptr_t* acc8, uintptr_t* acc32, uint32_t acc8_word_ct) {
   uint32_t widx;
-  uint32_t loader;
+  uintptr_t loader;
   for (widx = 0; widx < acc8_word_ct; widx++) {
     loader = *acc8++;
     *acc32 += (uint8_t)loader;
@@ -2274,7 +2274,7 @@ static inline void unroll_incr_8_32(const uintptr_t* acc8, uintptr_t* acc32, uin
 
 static inline void unroll_zero_incr_8_32(uintptr_t* acc8, uintptr_t* acc32, uint32_t acc8_word_ct) {
   uint32_t widx;
-  uint32_t loader;
+  uintptr_t loader;
   for (widx = 0; widx < acc8_word_ct; widx++) {
     loader = *acc8;
     *acc8++ = 0;
